Various import cleanup and rename util to visualization
